Enhance notifications architecture
